Added stack loading with files

+++ AED/TDAPila/TDAPila.c
@@ -7,45 +7,45 @@ void crearPila(tPila* pila) {
     pila -> tope = TAM_PILA;
 }
 
-int apilar(tPila* pila, const void* elem, size_t tamElem) {
+int apilar(tPila* pila, const void* elem, unsigned int tamElem) {
 
-    if (pila -> tope < (tamElem + sizeof(size_t)))
+    if (pila -> tope < (tamElem + sizeof(unsigned int)))
         return LLENO;
 
     pila -> tope -= tamElem;
     memcpy(pila -> pila + pila -> tope, elem, tamElem);
-    pila -> tope -= sizeof(size_t);
-    memcpy(pila -> pila + pila -> tope, &tamElem, sizeof(size_t));
+    pila -> tope -= sizeof(unsigned int);
+    memcpy(pila -> pila + pila -> tope, &tamElem, sizeof(unsigned int));
 
     return TODO_OK;
 }
 
-int verTope(const tPila* pila, void* elem, size_t tamElem) {
+int verTope(const tPila* pila, void* elem, unsigned int tamElem) {
 
-    size_t tamEnPila;
+    unsigned int tamEnPila;
 
     if (pila -> tope == TAM_PILA)
         return VACIO;
 
-    memcpy(&tamEnPila, pila -> pila + pila -> tope, sizeof(size_t));
+    memcpy(&tamEnPila, pila -> pila + pila -> tope, sizeof(unsigned int));
     // Va el minimo entre los dos para que contemple todos los largos del dato donde se guarda.
-    memcpy(elem, pila -> pila + pila -> tope + sizeof(size_t), MIN(tamElem, tamEnPila));
+    memcpy(elem, pila -> pila + pila -> tope + sizeof(unsigned int), MIN(tamElem, tamEnPila));
 
     return TODO_OK;
 }
 
-int desapilar(tPila* pila, void* elem, size_t tamElem) {
+int desapilar(tPila* pila, void* elem, unsigned int tamElem) {
 
-    size_t tamEnPila;
+    unsigned int tamEnPila;
 
     if (pila -> tope == TAM_PILA)
         return VACIO;
 
-    memcpy(&tamEnPila, pila -> pila + pila -> tope, sizeof(size_t));
-    pila -> tope += sizeof(size_t);
+    memcpy(&tamEnPila, pila -> pila + pila -> tope, sizeof(unsigned int));
+    pila -> tope += sizeof(unsigned int);
     // Va el minimo entre los dos para que contemple todos los largos del dato donde se guarda.
     memcpy(elem, pila -> pila + pila -> tope, MIN(tamElem, tamEnPila));
-    // Se corre el tamaño que ocupa en la pila, no importa la cantidad de bytes que se devuelvan.
+    // Se corre el tamaï¿½o que ocupa en la pila, no importa la cantidad de bytes que se devuelvan.
     pila -> tope += tamEnPila;
 
     return TODO_OK;
@@ -63,9 +63,9 @@ int pilaVacia(const tPila* pila) {
     return NO_VACIO;
 }
 
-int pilaLlena (const tPila* pila, size_t tamElem) {
+int pilaLlena (const tPila* pila, unsigned int tamElem) {
 
-    if (pila -> tope < (tamElem + sizeof(size_t)))
+    if (pila -> tope < (tamElem + sizeof(unsigned int)))
         return LLENO;
 
     return NO_LLENO;
